Integrate interrupt handling mechanism.
